afh_api/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
fonttools==4.57.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.3.2
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
